requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
lizard==1.17.10
pycparser==2.21
PyDriller==2.3
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pytz==2022.7
requests==2.28.1
smmap==5.0.0
types-pytz==2022.7.0.0
urllib3==1.26.13
wrapt==1.14.1